Share some utility operators
