+++ denotational/core/src/main/kotlin/io/github/oxidefrp/core/shared/StateScheduler.kt
@@ -5,7 +5,6 @@ import io.github.oxidefrp.core.EventStream
 import io.github.oxidefrp.core.Moment
 import io.github.oxidefrp.core.Signal
 import io.github.oxidefrp.core.mapNotNull
-import io.github.oxidefrp.core.orElse
 import io.github.oxidefrp.core.pullOf
 import io.github.oxidefrp.core.squashWith
